// web/src/api/generated/models/feedback-detail-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { FeedbackReplyResponse } from './feedback-reply-response';

export interface FeedbackDetailResponse {
    'id'?: number;
    'userId'?: number;
    'feedbackType'?: string;
    'title'?: string;
    'content'?: string;
    'attachment'?: string;
    'status'?: number;
    'statusDesc'?: string;
    'adminId'?: number;
    'processTime'?: string;
    'createTime'?: string;
    'updateTime'?: string;
    'replies'?: Array<FeedbackReplyResponse>;
}